break week component into smaller components, which fixes an issue with color change on conflicting courses
